port prototype tests to new utilities
